# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dashboard {
  activeCases: number;
  reviewedToday: number;
  urgentAlerts: number;
  accuracy: number;
  trend: number[];
}

export type MedicalCaseStatus = typeof MedicalCaseStatus[keyof typeof MedicalCaseStatus];


export const MedicalCaseStatus = {
  active: 'active',
  review: 'review',
  archived: 'archived',
} as const;

export type MedicalCasePriority = typeof MedicalCasePriority[keyof typeof MedicalCasePriority];


export const MedicalCasePriority = {
  normal: 'normal',
  urgent: 'urgent',
} as const;

export interface MedicalCase {
  id: number;
  patientLabel: string;
  title: string;
  specialty: string;
  status: MedicalCaseStatus;
  priority: MedicalCasePriority;
  updatedAt: string;
  confidence: number;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  sex?: string | null;
  tags?: string[];
}

export type CaseMessageRole = typeof CaseMessageRole[keyof typeof CaseMessageRole];


export const CaseMessageRole = {
  doctor: 'doctor',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface CaseMessage {
  id: number;
  role: CaseMessageRole;
  content: string;
  createdAt: string;
  citations?: string[];
}

export type FindingTone = typeof FindingTone[keyof typeof FindingTone];


export const FindingTone = {
  positive: 'positive',
  caution: 'caution',
  neutral: 'neutral',
  urgent: 'urgent',
} as const;

export interface Finding {
  label: string;
  value: string;
  tone: FindingTone;
}

export type MedicalCaseDetail = MedicalCase & {
  messages: CaseMessage[];
  summary: string;
  findings: Finding[];
};

export type CaseReportSectionsItem = {
  title: string;
  body: string;
};

export interface CaseReport {
  caseId: number;
  title: string;
  generatedAt: string;
  sections: CaseReportSectionsItem[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  analysis: 'analysis',
  report: 'report',
  upload: 'upload',
  alert: 'alert',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  detail: string;
  time: string;
}

export interface CaseInput {
  /** @minLength 1 */
  patientLabel: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  specialty: string;
  /**
     * @minimum 0
     * @maximum 120
     */
  age?: number;
  sex?: string;
  tags?: string[];
}

export type CaseUpdateStatus = typeof CaseUpdateStatus[keyof typeof CaseUpdateStatus];


export const CaseUpdateStatus = {
  active: 'active',
  review: 'review',
  archived: 'archived',
} as const;

export type CaseUpdatePriority = typeof CaseUpdatePriority[keyof typeof CaseUpdatePriority];


export const CaseUpdatePriority = {
  normal: 'normal',
  urgent: 'urgent',
} as const;

export interface CaseUpdate {
  /** @minLength 1 */
  title?: string;
  status?: CaseUpdateStatus;
  priority?: CaseUpdatePriority;
  tags?: string[];
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type ListCasesParams = {
status?: ListCasesStatus;
search?: string;
};

export type ListCasesStatus = typeof ListCasesStatus[keyof typeof ListCasesStatus];


export const ListCasesStatus = {
  active: 'active',
  review: 'review',
  archived: 'archived',
} as const;