# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseInput {
  /** @minLength 1 */
  patientLabel: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  specialty: string;
  /**
     * @minimum 0
     * @maximum 120
     */
  age?: number;
  sex?: string;
  tags?: string[];
}