# lib/api-zod/src/generated/types/caseMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseMessageRole } from './caseMessageRole';

export interface CaseMessage {
  id: number;
  role: CaseMessageRole;
  content: string;
  createdAt: string;
  citations?: string[];
}