# lib/api-zod/src/generated/types/caseReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseReportSectionsItem } from './caseReportSectionsItem';

export interface CaseReport {
  caseId: number;
  title: string;
  generatedAt: string;
  sections: CaseReportSectionsItem[];
}