# lib/api-zod/src/generated/types/caseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseUpdatePriority } from './caseUpdatePriority';
import type { CaseUpdateStatus } from './caseUpdateStatus';

export interface CaseUpdate {
  /** @minLength 1 */
  title?: string;
  status?: CaseUpdateStatus;
  priority?: CaseUpdatePriority;
  tags?: string[];
}