# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingTone } from './findingTone';

export interface Finding {
  label: string;
  value: string;
  tone: FindingTone;
}