# lib/api-zod/src/generated/types/medicalCaseDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseMessage } from './caseMessage';
import type { Finding } from './finding';
import type { MedicalCase } from './medicalCase';

export type MedicalCaseDetail = MedicalCase & {
  messages: CaseMessage[];
  summary: string;
  findings: Finding[];
};